refactor(frontend): group components/ by responsibility (#293)
The top level of components/ held 58 files with no organising principle.
Groups them per the issue's own proposal, as pure moves + import rewrites —
zero behaviour change.

  marketing/  HowItWorks, SignInModal, HeroCard (+test)
  graph/      KnowledgeGraph, 2D, 3D (+3 tests)
  chat/       ChatPanel, MarkdownChat, MermaidBlock, FunctionPlot,
              AIDisclaimerChip, ModelToggle, SharedContextToggle, SessionSummary

Top level: 58 -> 38 files. `ui/`, `flashcards/` and `screens/` are unchanged.

The proposal also listed study/ and social/ groups; neither had any top-level
members (those screens already live in screens/), so creating them would have
made empty directories.

Deliberately NOT done: renaming Gradebook/ to gradebook/. The proposal wanted
"the two Gradebook levels merged", but they are already split cleanly by
responsibility — components/Gradebook/ holds the parts, components/screens/
Gradebook/ holds the routes. The only remaining delta is letter case, and a
case-only directory rename is a genuine cross-platform hazard for the macOS
contributors on this repo. Not worth it for a cosmetic gain.

Moves were done with `git mv`, so history follows the files.

Four things had to move WITH the files, each of which would have failed
silently or confusingly otherwise:

- frontend/eslint.config.mjs — the testid-enforcement `files` array is keyed
  by path; stale entries would have silently stopped enforcing the rule on
  the exact files it was written to cover.
- frontend/eslint-suppressions.json — also path-keyed. Stale keys surfaced as
  22 lint errors, since the suppressions no longer matched any file.
- docs/frontend-testids.md — the surface table names owning files. This is
  the other half of the testid convention.
- HeroCard.test.tsx resolved src/ as `__dirname/..`, which silently became
  components/ once the file moved a level deeper. Now `../..`.

Also refreshed the live path references in SECURITY.md and the graph journey's
header comment. Historical records (docs/frontend-audit/, docs/decisions/,
docs/superpowers/plans/) are deliberately left pointing at the old paths —
they are dated snapshots of what was true when written, not live indexes.

part of #293
