+++ frontend/eslint-suppressions.json
@@ -50,35 +50,7 @@
       "count": 2
     }
   },
-  "src/components/KnowledgeGraph2D.tsx": {
-    "react-hooks/exhaustive-deps": {
-      "count": 1
-    },
-    "react-hooks/refs": {
-      "count": 12
-    }
-  },
-  "src/components/KnowledgeGraph3D.test.tsx": {
-    "@typescript-eslint/no-explicit-any": {
-      "count": 6
-    }
-  },
-  "src/components/ModelToggle.tsx": {
-    "react-hooks/set-state-in-effect": {
-      "count": 1
-    }
-  },
   "src/components/SessionFeedbackGlobal.tsx": {
-    "react-hooks/set-state-in-effect": {
-      "count": 1
-    }
-  },
-  "src/components/SharedContextToggle.tsx": {
-    "react-hooks/set-state-in-effect": {
-      "count": 1
-    }
-  },
-  "src/components/SignInModal.tsx": {
     "react-hooks/set-state-in-effect": {
       "count": 1
     }
@@ -86,6 +58,34 @@
   "src/components/TopNav.test.tsx": {
     "@typescript-eslint/no-explicit-any": {
       "count": 3
+    }
+  },
+  "src/components/chat/ModelToggle.tsx": {
+    "react-hooks/set-state-in-effect": {
+      "count": 1
+    }
+  },
+  "src/components/chat/SharedContextToggle.tsx": {
+    "react-hooks/set-state-in-effect": {
+      "count": 1
+    }
+  },
+  "src/components/graph/KnowledgeGraph2D.tsx": {
+    "react-hooks/exhaustive-deps": {
+      "count": 1
+    },
+    "react-hooks/refs": {
+      "count": 12
+    }
+  },
+  "src/components/graph/KnowledgeGraph3D.test.tsx": {
+    "@typescript-eslint/no-explicit-any": {
+      "count": 6
+    }
+  },
+  "src/components/marketing/SignInModal.tsx": {
+    "react-hooks/set-state-in-effect": {
+      "count": 1
     }
   },
   "src/components/screens/Admin.tsx": {
