+++ frontend/src/components/chat/ChatPanel.test.tsx
