+++ frontend/src/components/chat/FunctionPlot.tsx
