+++ frontend/src/components/chat/MermaidBlock.tsx
