+++ frontend/src/components/chat/SessionSummary.tsx
