+++ frontend/src/components/chat/SharedContextToggle.tsx
