+++ frontend/src/components/graph/KnowledgeGraph2D.testmode.test.tsx
