+++ frontend/src/components/graph/KnowledgeGraph2D.tsx
