+++ frontend/src/components/graph/KnowledgeGraph3D.test.tsx
