+++ frontend/src/components/graph/KnowledgeGraph3D.testmode.test.tsx
