+++ frontend/src/components/graph/KnowledgeGraph3D.tsx
