+++ frontend/src/components/marketing/HeroCard.tsx
