+++ frontend/src/components/screens/Learn.tsx
@@ -6,15 +6,15 @@ import { ChevronLeft } from "lucide-react";
 import { TopBar } from "../TopBar";
 import { Icon } from "../Icon";
 import { CustomSelect } from "../CustomSelect";
-import { ChatPanel, type ChatMsg } from "../ChatPanel";
+import { ChatPanel, type ChatMsg } from "../chat/ChatPanel";
 import { FullHeightScreen } from "../FullHeightScreen";
-import { SessionSummary } from "../SessionSummary";
-import { SharedContextToggle, useSharedContext } from "../SharedContextToggle";
-import { ModelToggle, useModelPref } from "../ModelToggle";
+import { SessionSummary } from "../chat/SessionSummary";
+import { SharedContextToggle, useSharedContext } from "../chat/SharedContextToggle";
+import { ModelToggle, useModelPref } from "../chat/ModelToggle";
 import { Toggle } from "@/components/ui";
 import { DisclaimerModal } from "../DisclaimerModal";
-import { AIDisclaimerChip } from "../AIDisclaimerChip";
-import { KnowledgeGraph } from "../KnowledgeGraph";
+import { AIDisclaimerChip } from "../chat/AIDisclaimerChip";
+import { KnowledgeGraph } from "../graph/KnowledgeGraph";
 import { useToast } from "../ToastProvider";
 import { useConfirm } from "@/lib/useConfirm";
 import { useIsMobile } from "@/lib/useIsMobile";
